bot-cinema/requirements.txt
discord.py
asyncpg
redis
python-dotenv
python-socketio[asyncio_client]
